requirements.txt
discord.py
playwright
pytz
pytesseract
Pillow
aiohttp
